memperbaiki sedikit data

+++ pertemuan10/index.php
@@ -38,7 +38,7 @@ $mahasiswa = query("SELECT * FROM mahasiswa");
                 <td><?= $i; ?></td>
                 <td>
                     <a href="">Ubah</a> |
-                    <a href="hapus.php?id=<?= $row["id"]; ?>">Hapus</a>
+                    <a href="hapus.php?id=<?= $row["id"]; ?>" onclick="return confirm ('yakin ingin menghapus data ini?')">Hapus</a>
                 </td>
                 <td><img src="img/<?= $row["gambar"]; ?>" alt="" width="50px"></td>
                 <td><?= $row["nim"] ?></td>
